Reorganize polza-cli: separate CLI and TUI into distinct folders
Split the monolithic structure into three main directories:
- cli/ - Traditional CLI interface (readline-based)
- tui/ - Modern TUI interface (React/Ink-based)
- shared/ - Common libraries and tools

Changes:
- Create separate cli/ and tui/ folders for better organization
- Move shared code (lib/, tools/) to shared/ directory
- Update all import paths to reflect new structure
- Update package.json entry points and scripts
- Remove obsolete files (src/index-old.js, README-old.md)
- Update test files to use new import paths
- Update README with new structure documentation

This fixes the Bun compatibility issue by clearly separating
the TUI dependencies (Ink/React) from the pure CLI code.

Fixes #87
