+++ polza-cli/shared/lib/settings-manager.js
