+++ polza-cli/tui/components/InputBar.jsx
