+++ polza-cli/tui/components/StatusBar.jsx
